+++ CHANGELOG.md
@@ -4,11 +4,17 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `sandstone/Input` props `backButtonAriaLabel` and `noBackButton` to customize a back button aria label and to not show a back button
+
 ### Changed
 
 - `sandstone/Input` layout for portrait mode
+- `sandstone/Input` to add a back button when `popupType` is `fullscreen`
 - `sandstone/Panels.Header` to always show back button
 - `sandstone/Scroller` scrollbar thumb to not get focus with touch when `focusableScrollbar` is `true` or `byEnter`
+- `sandstone/Scroller` and `sandstone/VirtualList` to show overscroll effect when flicking
 
 ### Fixed